feat(web): split dnd-kit out of main bundle, migrate UserMenu to Menu
Follow-up to the mobile-first meeting:

- Lazy-load BehaviorTracking inside the rewards route so dnd-kit only
  loads once a parent actually opens /rewards (not on initial page
  load).
- Lazy-load the crisis-list page the same way we did for the
  dashboard: the route file becomes a thin Suspense wrapper, the
  component body moves to crisis-list-page.tsx and imports
  dnd-kit/core + dnd-kit/sortable inside the lazy chunk.

Together these two splits move @dnd-kit/* (~46 kB raw / 15.5 kB gzip)
out of the main bundle into a shared sortable.esm chunk that only
loads on /rewards or /crisis-list. Main index bundle: 1399 kB -> 1355
kB (gzip 439 -> 426).

- Add a DropdownMenu wrapper (components/ui/dropdown-menu.tsx) around
  Base UI's Menu primitive and migrate UserMenu to use it. The
  primitive gives us real role=menu / role=menuitem semantics, arrow-
  key navigation between items, type-ahead, automatic focus-in on
  open and focus-return to the trigger on close, without hand-rolling
  any of it. Drops the manual open/setOpen state and role attributes
  I added last round.

Also removed a pre-existing English "More" sr-only string inside the
unused BreadcrumbEllipsis helper; its parent span is already
aria-hidden, so the sr-only text was unreachable anyway.
