+++ apps/web/src/components/ui/breadcrumb.tsx
@@ -110,9 +110,7 @@ function BreadcrumbEllipsis({
       )}
       {...props}
     >
-      <MoreHorizontalIcon
-      />
-      <span className="sr-only">More</span>
+      <MoreHorizontalIcon />
     </span>
   )
 }
