+++ apps/web/src/routes/_authenticated.tsx
@@ -7,14 +7,16 @@ import {
 } from "@tanstack/react-router";
 import { useTranslation } from "react-i18next";
 import { Heart, LogOut, LifeBuoy, ChevronDown, Menu, Lock } from "lucide-react";
-import { useState } from "react";
 import { Button } from "@/components/ui/button";
 import { Separator } from "@/components/ui/separator";
 import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@/components/ui/popover";
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuLabel,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger,
+} from "@/components/ui/dropdown-menu";
 import {
   Sidebar,
   SidebarContent,
@@ -217,7 +219,6 @@ function AppSidebar() {
 function UserMenu() {
   const { t } = useTranslation();
   const session = useSession();
-  const [open, setOpen] = useState(false);
   const { openKoe, available: koeAvailable } = useKoeTrigger();
 
   const user = session.data?.user;
@@ -237,8 +238,8 @@ function UserMenu() {
   };
 
   return (
-    <Popover open={open} onOpenChange={setOpen}>
-      <PopoverTrigger
+    <DropdownMenu>
+      <DropdownMenuTrigger
         render={
           <SidebarMenuButton
             size="lg"
@@ -261,15 +262,9 @@ function UserMenu() {
           )}
         </span>
         <ChevronDown className="h-3.5 w-3.5 text-muted-foreground group-data-[collapsible=icon]:hidden" />
-      </PopoverTrigger>
-      <PopoverContent
-        align="end"
-        side="top"
-        role="menu"
-        aria-label={t("nav.userMenu")}
-        className="w-60 gap-0 p-1"
-      >
-        <div className="flex flex-col gap-0.5 px-3 py-2">
+      </DropdownMenuTrigger>
+      <DropdownMenuContent align="end" side="top" className="w-60">
+        <DropdownMenuLabel>
           <span className="truncate text-sm font-medium text-foreground">
             {user.name}
           </span>
@@ -278,48 +273,24 @@ function UserMenu() {
               {user.email}
             </span>
           )}
-        </div>
-        <Separator className="my-1" />
+        </DropdownMenuLabel>
+        <DropdownMenuSeparator />
         {koeAvailable && (
-          <button
-            type="button"
-            role="menuitem"
-            onClick={() => {
-              setOpen(false);
-              openKoe();
-            }}
-            className="flex w-full items-center gap-2 rounded-md px-2 py-2 text-sm hover:bg-accent focus-visible:bg-accent focus-visible:outline-none"
-          >
+          <DropdownMenuItem onClick={openKoe}>
             <LifeBuoy className="h-4 w-4 text-muted-foreground" aria-hidden="true" />
             {t("nav.support")}
-          </button>
+          </DropdownMenuItem>
         )}
-        <button
-          type="button"
-          role="menuitem"
-          onClick={() => {
-            setOpen(false);
-            useUiStore.getState().lock();
-          }}
-          className="flex w-full items-center gap-2 rounded-md px-2 py-2 text-sm hover:bg-accent focus-visible:bg-accent focus-visible:outline-none"
-        >
+        <DropdownMenuItem onClick={() => useUiStore.getState().lock()}>
           <Lock className="h-4 w-4 text-muted-foreground" aria-hidden="true" />
           {t("nav.lock")}
-        </button>
-        <button
-          type="button"
-          role="menuitem"
-          onClick={() => {
-            setOpen(false);
-            handleSignOut();
-          }}
-          className="flex w-full items-center gap-2 rounded-md px-2 py-2 text-sm hover:bg-accent focus-visible:bg-accent focus-visible:outline-none"
-        >
+        </DropdownMenuItem>
+        <DropdownMenuItem onClick={handleSignOut}>
           <LogOut className="h-4 w-4 text-muted-foreground" aria-hidden="true" />
           {t("nav.logout")}
-        </button>
-      </PopoverContent>
-    </Popover>
+        </DropdownMenuItem>
+      </DropdownMenuContent>
+    </DropdownMenu>
   );
 }
